Fixing new staticcheck findings
