+++ internal/driver/glfw/driver_desktop.go
@@ -9,7 +9,6 @@ import (
 	"os/signal"
 	"runtime"
 	"syscall"
-	"time"
 
 	"fyne.io/fyne/v2/canvas"
 	"fyne.io/fyne/v2/internal/painter"
@@ -20,8 +19,6 @@ import (
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/theme"
 )
-
-const desktopDefaultDoubleTapDelay = 300 * time.Millisecond
 
 var (
 	systrayIcon    fyne.Resource
