+++ internal/driver/glfw/window_notdarwin.go
@@ -2,6 +2,8 @@
 
 package glfw
 
+const desktopDefaultDoubleTapDelay = 300 * time.Millisecond
+
 func (w *window) doSetFullScreen(full bool) {
 	monitor := w.getMonitorForWindow()
 	mode := monitor.GetVideoMode()
